requirements.txt
Django>=2.0
django-braces>=1.11.0
django-crispy-forms>=1.6.1
django-extensions>=1.7.7
python-jose>=3.0.0
warrant-lite>=1.0.2
